+++ src/app/[variants]/(main)/(mobile)/me/settings/features/useCategory.tsx
@@ -2,14 +2,11 @@ import { Bot, Brain, Info, Mic2, Settings2, Sparkles } from 'lucide-react';
 import { useTranslation } from 'react-i18next';
 import { useNavigate } from 'react-router-dom';
 import { CellProps } from '@/components/Cell';
-import { isDeprecatedEdition } from '@/const/version';
 import { SettingsTabs } from '@/store/global/initialState';
-import { featureFlagsSelectors, useServerConfigStore } from '@/store/serverConfig';
 
 export const useCategory = () => {
   const navigate = useNavigate();
   const { t } = useTranslation('setting');
-  const { showLLM } = useServerConfigStore(featureFlagsSelectors);
 
   const items: CellProps[] = [
     {
@@ -22,18 +19,11 @@ export const useCategory = () => {
       key: SettingsTabs.SystemAgent,
       label: t('tab.system-agent'),
     },
-    showLLM &&
-      (isDeprecatedEdition
-        ? {
-            icon: Brain,
-            key: SettingsTabs.LLM,
-            label: t('tab.llm'),
-          }
-        : {
-            icon: Brain,
-            key: SettingsTabs.Provider,
-            label: t('tab.provider'),
-          }),
+    {
+      icon: Brain,
+      key: SettingsTabs.Provider,
+      label: t('tab.provider'),
+    },
     { icon: Mic2, key: SettingsTabs.TTS, label: t('tab.tts') },
     {
       icon: Bot,