+++ src/app/[variants]/(main)/profile/hooks/useCategory.tsx
@@ -4,8 +4,6 @@ import { Link } from 'react-router-dom';
 import { useTranslation } from 'react-i18next';
 
 import type { MenuProps } from '@/components/Menu';
-import { enableAuth } from '@/const/auth';
-import { isDeprecatedEdition } from '@/const/version';
 import { ProfileTabs } from '@/store/global/initialState';
 import { featureFlagsSelectors, useServerConfigStore } from '@/store/serverConfig';
 import { useUserStore } from '@/store/user';
@@ -26,7 +24,6 @@ export const useCategory = () => {
         </Link>
       ),
     },
-    enableAuth &&
     isLoginWithClerk && {
       icon: <Icon icon={ShieldCheck} />,
       key: ProfileTabs.Security,
@@ -36,7 +33,7 @@ export const useCategory = () => {
         </Link>
       ),
     },
-    !isDeprecatedEdition && {
+    {
       icon: <Icon icon={ChartColumnBigIcon} />,
       key: ProfileTabs.Stats,
       label: (