+++ src/features/Conversation/Messages/Group/Group.tsx
@@ -4,6 +4,10 @@ import isEqual from 'fast-deep-equal';
 import { memo, useMemo } from 'react';
 import { Flexbox } from 'react-layout-kit';
 
+import { CollapsedMessage } from '@/features/Conversation/Messages/Group/CollapsedMessage';
+import { useChatStore } from '@/store/chat';
+import { messageStateSelectors } from '@/store/chat/slices/message/selectors';
+
 import { GroupMessageContext } from './GroupContext';
 import GroupItem from './GroupItem';
 
@@ -19,18 +23,28 @@ const useStyles = createStyles(({ css }) => {
 
 interface GroupChildrenProps {
   blocks: AssistantContentBlock[];
+  content?: string;
   contentId?: string;
   disableEditing?: boolean;
   id: string;
   messageIndex: number;
 }
 
-const GroupChildren = memo<GroupChildrenProps>(
-  ({ blocks, contentId, disableEditing, messageIndex, id }) => {
+const Group = memo<GroupChildrenProps>(
+  ({ blocks, contentId, disableEditing, messageIndex, id, content }) => {
     const { styles } = useStyles();
-
+    const [isCollapsed] = useChatStore((s) => [messageStateSelectors.isMessageCollapsed(id)(s)]);
     const contextValue = useMemo(() => ({ assistantGroupId: id }), [id]);
 
+    if (isCollapsed) {
+      return (
+        content && (
+          <Flexbox>
+            <CollapsedMessage content={content} id={id} />
+          </Flexbox>
+        )
+      );
+    }
     return (
       <GroupMessageContext value={contextValue}>
         <Flexbox className={styles.container} gap={8}>
@@ -53,4 +67,4 @@ const GroupChildren = memo<GroupChildrenProps>(
   isEqual,
 );
 
-export default GroupChildren;
+export default Group;